% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilespec.R
\name{as.tile_source}
\alias{as.tile_source}
\alias{is.tile_source}
\alias{source_from_url_format}
\title{Tile Sources}
\usage{
as.tile_source(x, ...)

is.tile_source(x)

source_from_url_format(url_format, max_zoom = tile.maxzoom.default(),
  min_zoom = 0, attribution = NULL, ...)
}
\arguments{
\item{x}{An object (usually a name or string format) with which to create a tile source}

\item{...}{Arguments passed to other methods}
}
\value{
An object of class 'tile_source'
}
\description{
Tile sources define where rosm looks for map tiles. There are a number of
built-in types (\link{osm.types}), or they can be created using \code{as.tile_source()},
registered using \link{register_tile_source} for easy access, or passed directly to
the \link{osm.plot} family of methods.
}
\details{
Passing a name from \link{osm.types} will return that tile source; passing a name from
\link{register_tile_source} will return that tile source, and passing a URL format in the form
\code{https://tiles.wmflabs.org/bw-mapnik/${z}/${x}/${y}.png} will create a new tile source.
Old style function names in the form tile.url.TYPE are still supported but are deprecated.
}
\examples{
as.tile_source("osm")

}

