% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bingmaps.R
\name{bmaps.plot}
\alias{bmaps.plot}
\title{Plot Bing Maps}
\usage{
bmaps.plot(bbox, bingtype = "Aerial", key = NULL, ...)
}
\arguments{
\item{bbox}{}

\item{bingtype}{}

\item{key}{}

\item{...}{}
}
\description{
Plot Bing Maps
}

