% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bingmaps.R
\name{bmaps.types}
\alias{bmaps.types}
\title{List types of Bing Maps}
\usage{
bmaps.types()
}
\value{
A list of valid bing map types
}
\description{
List types of Bing Maps
}
\examples{
bmaps.types()
}

