% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmapplot.R
\name{gmap.plot}
\alias{gmap.plot}
\title{Plot Google Maps}
\usage{
gmap.plot(bbox, maptype = "satellite", forcedownload = FALSE,
  cachedir = NULL, res = 150, project = TRUE, key = NULL, ...)
}
\arguments{
\item{bbox}{}

\item{maptype}{}

\item{forcedownload}{}

\item{cachedir}{}

\item{res}{}

\item{key}{}

\item{...}{}

\item{zoomin}{}

\item{zoom}{}

\item{epsg}{}
}
\description{
Plot google maps.
}

