% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{has_internet}
\alias{has_internet}
\title{Check for Internet}
\usage{
has_internet()
}
\value{
TRUE if the internet is available, false otherwise
}
\description{
Check for Internet
}
\examples{
has_internet()

}
