% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlays.R
\name{osm.lines}
\alias{osm.lines}
\title{Overlay lines on an OSM plot}
\usage{
osm.lines(x, y = NULL, epsg = 4326, toepsg = 3857, ...)
}
\arguments{
\item{x}{X coordinate vector or object as parsed by \code{xy.coords}}

\item{y}{Y coordinate vector}

\item{epsg}{EPSG code of the supplied coordinates}

\item{toepsg}{EPSG code of the projected coordinates to be ploted}

\item{...}{Args passed on to \code{lines}}
}
\description{
Plot lines on a plot created by \link{osm.plot}. This is a simple wrapper around
\code{points()}.
}
\examples{
\dontshow{if (identical(Sys.getenv("R_PRETTYMAPR_HAS_API_KEY"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
library(rosm)
library(prettymapr)
locs <- geocode(c("wolfville, ns", "kentville, ns", "halifax, ns"))
prettymap({
  osm.plot(searchbbox("nova scotia"))
  osm.lines(locs$lon, locs$lat, lwd = 2)
})
}
\dontshow{\}) # examplesIf}
}
