% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmraster.R
\name{osm.proj}
\alias{osm.proj}
\title{Project an OSM RasterStack}
\usage{
osm.proj(osm.raster, projection, crop.bbox = NULL)
}
\arguments{
\item{osm.raster}{A RasterStack}

\item{projection}{A projection as generated by \code{Spatial*@proj4String} or \code{sp::CRS()}.}

\item{crop.bbox}{(optional) a bbox to crop the image to (in projected coordinates)}
}
\description{
Project an OSM RasterStack
}
\examples{
\donttest{
library(cartography)
library(raster)
library(prettymapr)

data(nuts2006)
spdf <- nuts0.spdf[nuts0.spdf$id=="DE",]
spdf2 <- spTransform(spdf, CRSobj = CRS("+init=epsg:4326"))
x <- osm.raster(bbox(spdf2), type="thunderforestlandscape")
x2 <- osm.proj(x, spdf@proj4string, crop.bbox=bbox(spdf))
osm.proj.plot(x2)

ns <- makebbox(47.2, -59.7, 43.3, -66.4)
x <- osm.raster(ns)
osm.proj.plot(osm.proj(x, CRS("+init=epsg:26920")))
}
}

