% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmraster.R
\name{osm.raster.plot}
\alias{osm.raster.plot}
\title{Plot an OSM RasterStack
Plot an OSM RasterStack as generated by \link{osm.raster}.
This function is a thin wrapper to \code{raster::plotRGB()}.}
\usage{
osm.raster.plot(rgbraster, ...)
}
\arguments{
\item{rgbraster}{An RGB RasterStack as generated by \link{osm.raster}}

\item{...}{Arguments to be passed on to \code{raster::plotRGB()}.}
}
\description{
Plot an OSM RasterStack
Plot an OSM RasterStack as generated by \link{osm.raster}.
This function is a thin wrapper to \code{raster::plotRGB()}.
}

