% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmraster.R
\name{osm.raster}
\alias{osm.raster}
\title{Get Open Street Map Tiles As RasterStack}
\usage{
osm.raster(bbox, zoomin = 0, zoom = NULL, type = "osm",
  forcedownload = FALSE, cachedir = NULL)
}
\arguments{
\item{bbox}{A bounding box as generated by \code{sp::bbox()} or \code{prettymapr::searchbbox()}.
Must be in lon/lat (epsg:4326)!}

\item{zoomin}{The amount by which to adjust the automatically calculated zoom (or
manually specified if the \code{zoom} parameter is passed). Use +1 to zoom in, or -1 to zoom out.}

\item{zoom}{Manually specify the zoom level (not reccomended; adjust \code{zoomin} or
\code{res} instead.}

\item{type}{A map type; one of that returned by \link{osm.types}. User defined types are possible
by defining \code{tile.url.TYPENAME <- function(xtile, ytile, zoom){}} and passing TYPENAME
as the \code{type} argument.}

\item{forcedownload}{\code{TRUE} if cached tiles should be re-downloaded. Useful if
some tiles are corrupted.}

\item{cachedir}{The directory in which tiles should be cached. Defaults to \code{getwd()/rosm.cache}.}
}
\value{
A RasterStack of the fused tiles is returned.
}
\description{
Get Open Street Map tiles as RasterStack object (requires package
\code{raster} to be installed) with a Spherical Mercator projection
(\href{https://en.wikipedia.org/wiki/Web_Mercator}{epsg:3857}).
}
\details{
Tiles are fused and projected
}
\examples{
\donttest{
library(cartography)
library(raster)
library(prettymapr)
data(nuts2006)
spdf <- nuts0.spdf[nuts0.spdf$id=="DE",]
spdf2 <- spTransform(spdf, CRSobj = CRS("+init=epsg:4326"))
x <- osm.raster(bbox(spdf2), type="thunderforestlandscape")

ns <- makebbox(47.2, -59.7, 43.3, -66.4)
x <- osm.raster(ns)
}
}

