% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{osm.text}
\alias{osm.text}
\title{Overlay text on an OSM plot}
\usage{
osm.text(x, y = NULL, labels = seq_along(x), epsg = 4326, toepsg = 3857, ...)
}
\arguments{
\item{x}{X coordinate vector or object as parsed by \code{xy.coords}}

\item{y}{Y coordinate vector}

\item{labels}{A character vector or expression specifying the text to be
written.}

\item{epsg}{EPSG code of the supplied coordinates}

\item{toepsg}{EPSG code of the projected coordinates to be ploted}

\item{...}{Args passed on to \code{text()}}
}
\description{
Plot text on a plot created by \link{osm.plot}.
}
