% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilespec.R
\name{osm.types}
\alias{osm.types}
\title{Get List of Valid Tile Sources}
\usage{
osm.types()
}
\value{
A character vector of valid \code{type} parameters.
}
\description{
Get List of Valid Tile Sources
}
\examples{
osm.types()

}

