% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachedir.R
\name{set_default_cachedir}
\alias{set_default_cachedir}
\alias{get_default_cachedir}
\title{Set/Get the Default Tile Cache Location}
\usage{
set_default_cachedir(cachedir)

get_default_cachedir()
}
\arguments{
\item{cachedir}{A path to use as the cache directory (relative to the working directory).
Use NULL to reset to the default.}
}
\value{
The previous cache directory, invisibly.
}
\description{
The default tile cache location is the "rosm.cache" folder in the current
working directory, but for a variety of reasons it may be desirable to
use one cache directory for all calls in a script. This must be called
every time the namespace is loaded.
}
\examples{
set_default_cachedir(tempfile())
get_default_cachedir()
(set_default_cachedir(NULL))

}
