% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googlestatic.R
\name{gmap.plot}
\alias{gmap.plot}
\title{Plot Google Maps}
\usage{
gmap.plot(bbox, zoomin = 0, zoom = NULL, maptype = "satellite",
  forcedownload = FALSE, cachedir = NULL, res = 150, epsg = 3857,
  key = NULL, ...)
}
\arguments{
\item{bbox}{}

\item{zoomin}{}

\item{zoom}{}

\item{maptype}{}

\item{forcedownload}{}

\item{cachedir}{}

\item{res}{}

\item{epsg}{}

\item{key}{}

\item{...}{}
}
\description{
Plot google maps.
}

