% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_internet}
\alias{has_internet}
\title{Check for Internet}
\usage{
has_internet()
}
\value{
TRUE if the internet is available, false otherwise
}
\description{
Used to skip tests and examples for this package when offline.
}
\examples{
has_internet()

}
