% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tileplot.R
\name{osm.plot}
\alias{osm.plot}
\title{Plot Open Street Map Tiles}
\usage{
osm.plot(bbox, zoomin = 0, zoom = NULL, type = "osm",
  forcedownload = FALSE, stoponlargerequest = TRUE, fusetiles = TRUE,
  cachedir = NULL, res = 150, epsg = 3857, ...)
}
\arguments{
\item{bbox}{}

\item{zoomin}{}

\item{zoom}{}

\item{type}{}

\item{forcedownload}{}

\item{stoponlargerequest}{}

\item{fusetiles}{}

\item{cachedir}{}

\item{res}{}

\item{epsg}{}

\item{...}{}
}
\description{
Plot open street map tiles.
}
\examples{
\donttest{
library(prettymapr)
ns <- searchbbox("nova scotia")
osm.plot(ns)
osm.plot(ns, zoomin=1)
osm.plot(ns, type="mapquestsat")
prettymap(osm.plot(ns), scale.style="ticks", scale.tick.cex=0)
}
}

