% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osmplot.R
\name{osm.types}
\alias{osm.types}
\title{Get List of Valid Types}
\usage{
osm.types()
}
\value{
A character vector of valid \code{type} parameters.
}
\description{
Get List of Valid Types
}
\examples{
osm.types()
}

