% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{osm_url_spec}
\alias{osm_url_spec}
\alias{as_osm_url_spec}
\title{Tile URL Specification}
\usage{
osm_url_spec(
  server_url = "https://tile.openstreetmap.org/${z}/${x}/${y}.png",
  block_size = c(256, 256),
  min_zoom = 0,
  max_zoom = 18
)

as_osm_url_spec(x, ...)
}
\arguments{
\item{server_url}{A url using `${x}`, `${y}`, and `${z}` for the x, y, and
zoom level to be replaced. This can be any URL; non-URLs are assumed to be
local file paths relative to the current working directory at the time of
the download.}

\item{block_size}{The pixel size of each image}

\item{min_zoom, max_zoom}{The min/max zoom that this tile specification can handle}

\item{x}{An object to convert to an osm_url_spec}

\item{...}{Passed to S3 methods}
}
\value{
An object of class osm_url_spec.
}
\description{
Tile URL Specification
}
\examples{
osm_url_spec()

}
